requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.2
matplotlib>=3.7.0
python-dotenv>=1.0.0
pytz>=2023.3
